fileLoad/requirements.txt
google-cloud-aiplatform >= 1.31.0 
langchain >= 0.0.201
pdf2image >= 1.16.3

tensorflow_hub >= 0.13.0 
functions-framework >= 3.3.0
cloudevents >= 1.9.0

tensorflow
tensorflow-hub
# tensorflow-text
tensorrt

# unstructured >= 0.7.5
# pytesseract >= 0.3.10
# pdfminer.six >= 20221105